Remove obj client interface
