+++ etc/helm/examples/local-dev-values.yaml
@@ -18,7 +18,6 @@ pachd:
   activateEnterprise: false
   # enterpriseLicenseKey: ""
   storage:
-    gocdkEnabled: true
     storageURL: s3://pachyderm-test?endpoint=minio.default.svc.cluster.local:9000&disableSSL=true&region=dummy-region
     backend: AMAZON
     amazon:
