+++ etc/helm/pachyderm/templates/pachd/deployment.yaml
@@ -53,9 +53,7 @@ spec:
         - --mode
         - "$(MODE)"
         env:
-        - name: GOCDK_ENABLED
-          value: {{ .Values.pachd.storage.gocdkEnabled | quote }}
-        {{- if and (ne .Values.deployTarget "LOCAL") .Values.pachd.storage.gocdkEnabled  }}
+        {{- if ne .Values.deployTarget "LOCAL" }}
         - name: STORAGE_URL
           value: {{ required "storage URL required" .Values.pachd.storage.storageURL | quote }}
         {{- end }}
