+++ etc/helm/pachyderm/templates/pachw/deployment.yaml
@@ -56,9 +56,7 @@ spec:
         env:
         - name: IS_PACHW
           value: "true"
-        - name: GOCDK_ENABLED
-          value: {{ .Values.pachd.storage.gocdkEnabled | quote }}
-        {{- if and (ne .Values.deployTarget "LOCAL") .Values.pachd.storage.gocdkEnabled  }}
+        {{- if ne .Values.deployTarget "LOCAL" }}
         - name: STORAGE_URL
           value: {{ required "storage URL required" .Values.pachd.storage.storageURL | quote }}
         {{- end }}
