+++ etc/helm/pachyderm/values.yaml
@@ -924,16 +924,11 @@ pachd:
     name: "pachyderm" #TODO Set default in helpers / Wire up in templates
   storage:
     # backend configures the storage backend to use.  It must be one
-    # of GOOGLE, AMAZON, MINIO, MICROSOFT or LOCAL. This is set automatically
-    # if deployTarget is GOOGLE, AMAZON, MICROSOFT, or LOCAL
-    backend: ""
-    # If 'gocdkEnabled' is true, 'storageURL' is used to configure the backend for object storage.
-    # The 'backend' configuration options should be used in conjunction with 'storageURL' to
-    # pass in secrets and identity information.
-    # When using gocdkEnabled, 'backend' cannot be 'MINIO'.
+    # of GOOGLE, AMAZON, MICROSOFT or LOCAL. This is set automatically
+    # if deployTarget is GOOGLE, AMAZON, MICROSOFT, or LOCAL.
     # Users who desire to use minio as their storage backend should set the 'backend' to 'AMAZON' instead
     # and set 'storageURL' to the s3 compatible URL representation for their minio endpoint.
-    gocdkEnabled: true
+    backend: ""
     # storageURL is a URL endpoint for object storage such as "s3://myamazonbucket" or "gcs://mygooglebucket.
     # Depending on the storage provider, some configuration options may also be passed via URL query parameters.
     storageURL: ""
