+++ etc/testing/circle/helm-values.yaml
@@ -13,7 +13,6 @@ pachd:
     storage:
         backend: AMAZON
         amazon:
-            gocdkEnabled: "true"
             storageURL: "s3://pachyderm-test?endpoint=minio.default.svc.cluster.local:9000&disableSSL=true&region=dummy-region"
             id: "minioadmin"
             secret: "minioadmin"
