+++ src/server/pps/server/worker_rc.go
@@ -145,9 +145,6 @@ func (kd *kubeDriver) workerPodSpec(ctx context.Context, options *workerOptions,
 		Name:  "STORAGE_URL",
 		Value: kd.config.StorageURL,
 	}, {
-		Name:  "GOCDK_ENABLED",
-		Value: strconv.FormatBool(kd.config.GoCDKEnabled),
-	}, {
 		Name:  "POSTGRES_USER",
 		Value: kd.config.PostgresUser,
 	}, {
