+++ src/server/worker/pipeline/transform/transform_test.go
@@ -41,7 +41,7 @@ func setupPachAndWorker(ctx context.Context, t *testing.T, dbConfig pachconfig.C
 
 	// Set env vars that the object storage layer expects in the env
 	// This is global but it should be fine because all tests use the same value.
-	require.NoError(t, os.Setenv(obj.StorageBackendEnvVar, obj.Local))
+	require.NoError(t, os.Setenv("STORAGE_BACKEND", obj.Local))
 	require.NoError(t, os.MkdirAll(env.ServiceEnv.Config().StorageRoot, 0777))
 
 	// Set up the input repo and branch
